allow affinity mask wider than 32 processors
the affinity cli option was parsed as int so a mask with any bit above 32
could not be passed at all even though the job stores it as IntPtr and
FixAffinity already handles a 64 bit mask

parsing it as long also means the perfonar model must hold long or the
value gets truncated on the way out

+++ src/BenchmarkDotNet/ConsoleArguments/CommandLineOptions.cs
@@ -64,7 +64,7 @@ namespace BenchmarkDotNet.ConsoleArguments
         public OutlierMode Outliers { get; set; }
 
         [Option("affinity", Required = false, HelpText = "Affinity mask to set for the benchmark process")]
-        public int? Affinity { get; set; }
+        public long? Affinity { get; set; }
 
         [Option("allStats", Required = false, Default = false, HelpText = "Displays all statistics (min, max & more)")]
         public bool DisplayAllStatistics { get; set; }
