+++ src/BenchmarkDotNet/Models/BdnEnvironment.cs
@@ -8,5 +8,5 @@ internal class BdnEnvironment : EnvironmentInfo
 {
     public RuntimeMoniker? Runtime { get; set; }
     public Jit? Jit { get; set; }
-    public int? Affinity { get; set; }
+    public long? Affinity { get; set; }
 }
